separating cookies function to decrease bundle weight
