+++ src/components/popup/details/details.test.js
@@ -5,6 +5,7 @@ import Store from '../../../lib/store';
 import purposesStyle from './purposes/purposes';
 
 import Details from './details';
+import {CMP_ID} from "../../../lib/init";
 
 describe('Details', () => {
 	let scratch;
@@ -15,6 +16,7 @@ describe('Details', () => {
 
 	it('should render with purpose panel initially', () => {
 		const store = new Store({
+			cmpId: CMP_ID,
 			vendorList: {
 				purposes: [
 					{ id: 1, name: 'Purpose 1' },
@@ -30,6 +32,7 @@ describe('Details', () => {
 
 	it('should switch between panel states', () => {
 		const store = new Store({
+			cmpId: CMP_ID,
 			vendorList: {
 				purposes: [
 					{ id: 1, name: 'Purpose 1' },
