+++ src/components/popup/details/purposes/purposes.jsx
@@ -114,7 +114,7 @@ export default class Purposes extends Component {
 			specialPurposes,
 			features,
 			specialFeatures,
-			persistedConsentData,
+			persistedConsentData = {},
 			initialVendorsRejection
 		} = props;
