+++ src/docs/assets/complete.html
@@ -84,6 +84,7 @@
 <script>
 	window.__cmp = {config: {logging: 'debug'}};
 </script>
-<script src="../cmp.complete.bundle.js"></script>
+<script src="../cmp.bundle.js"></script>
+<script src="../cookie.bundle.js"></script>
 </body>
 </html>
