+++ src/lib/cookie/cookie.test.js
@@ -1,15 +1,12 @@
 /* eslint-disable max-nested-callbacks */
-
 import { expect } from 'chai';
 import {
-	writeCookie,
 	decodeConsentData,
 	encodeConsentData,
 	writeConsentCookie,
 	readConsentCookie,
 	convertVendorsToRanges,
 	encodeVendorConsentData,
-	decodeVendorConsentData,
 	CONSENT_COOKIE
 } from './cookie';
 import {
@@ -23,24 +20,65 @@ import {
 	VENDOR_LIST
 } from "../../../test/constants";
 import { GVL, TCModel } from "@iabtcf/core";
-import Promise from "promise-polyfill";
+import {decodeVendorCookieValue, decodeBitsToIds, decodePublisherCookieValue} from "./cookieDecodeHelpers";
+import {encodePublisherCookieValue, encodeVendorCookieValue} from "./cookieEncodeHelpers";
 
-jest.mock('../portal');
-const mockPortal = require('../portal');
+function decodeVendorConsentData(cookieValue) {
+	const {
+		cookieVersion,
+		cmpId,
+		cmpVersion,
+		consentScreen,
+		consentLanguage,
+		vendorListVersion,
+		purposeIdBitString,
+		maxVendorId,
+		created,
+		lastUpdated,
+		isRange,
+		defaultConsent,
+		vendorIdBitString,
+		vendorRangeList
+	} = decodeVendorCookieValue(cookieValue);
+
+	const cookieData = {
+		cookieVersion,
+		cmpId,
+		cmpVersion,
+		consentScreen,
+		consentLanguage,
+		vendorListVersion,
+		selectedPurposeIds: decodeBitsToIds(purposeIdBitString),
+		maxVendorId,
+		created,
+		lastUpdated
+	};
+
+	if (isRange) {
+		const idMap = vendorRangeList.reduce((acc, {isRange, startVendorId, endVendorId}) => {
+			const lastVendorId = isRange ? endVendorId : startVendorId;
+			for (let i = startVendorId; i <= lastVendorId; i++) {
+				acc[i] = true;
+			}
+			return acc;
+		}, {});
+
+		cookieData.selectedVendorIds = new Set();
+		for (let i = 0; i <= maxVendorId; i++) {
+			if ((defaultConsent && !idMap[i]) ||
+				(!defaultConsent && idMap[i])) {
+				cookieData.selectedVendorIds.add(i);
+			}
+		}
+	}
+	else {
+		cookieData.selectedVendorIds = decodeBitsToIds(vendorIdBitString);
+	}
+
+	return cookieData;
+}
 
 describe('cookie', () => {
-	beforeEach(() => {
-		// Remove all cookies
-		const value = document.cookie.split(';');
-		value.forEach(cookie => {
-			const parts = cookie.trim().split('=');
-			if (parts.length === 2) {
-				writeCookie(parts[0], '', 0);
-			}
-		});
-		mockPortal.sendPortalCommand = jest.fn().mockImplementation(() => Promise.resolve());
-	});
-
 	it('encodes and decodes consent cookie object back to original value', (done) => {
 		const tcModel = new TCModel();
 		tcModel.cmpId = 280;
@@ -225,5 +263,131 @@ describe('cookie', () => {
 				endVendorId: undefined
 			}
 		]);
+	});
+});
+
+describe('cookie common', () => {
+	it('encodes and decodes the vendor cookie value with ranges back to original value', () => {
+
+		const aDate = new Date('2018-07-15 PDT');
+
+		const consentData = {
+			cookieVersion: 1,
+			created: aDate,
+			lastUpdated: aDate,
+			cmpId: 1,
+			cmpVersion: 1,
+			consentScreen: 1,
+			consentLanguage: 'EN',
+			vendorListVersion: 1,
+			purposeIdBitString: '111000001010101010001101',
+			maxVendorId: 5,
+			isRange: true,
+			defaultConsent: false,
+			numEntries: 2,
+			vendorRangeList: [
+				{
+					isRange: true,
+					startVendorId: 2,
+					endVendorId: 4
+				},
+				{
+					isRange: false,
+					startVendorId: 1
+				}
+			]
+		};
+
+		const bitString = encodeVendorCookieValue(consentData);
+		const decoded = decodeVendorCookieValue(bitString);
+
+		expect(decoded).to.deep.equal(consentData);
+	});
+
+	it('encodes and decodes the vendor cookie value with range ranges back to original value', () => {
+
+		const aDate = new Date('2018-07-15 PDT');
+
+		const consentData = {
+			cookieVersion: 1,
+			created: aDate,
+			lastUpdated: aDate,
+			cmpId: 1,
+			cmpVersion: 1,
+			consentScreen: 1,
+			consentLanguage: 'EN',
+			vendorListVersion: 1,
+			purposeIdBitString: '111000001010101010001101',
+			maxVendorId: 5,
+			isRange: true,
+			defaultConsent: false,
+			numEntries: 2,
+			vendorRangeList: [
+				{
+					isRange: false,
+					startVendorId: 2
+				},
+				{
+					isRange: false,
+					startVendorId: 1
+				}
+			]
+		};
+
+		const bitString = encodeVendorCookieValue(consentData);
+		const decoded = decodeVendorCookieValue(bitString);
+
+		expect(decoded).to.deep.equal(consentData);
+	});
+
+	it('encodes and decodes the vendor cookie value without ranges back to original value', () => {
+
+		const aDate = new Date('2018-07-15 PDT');
+
+		const consentData = {
+			cookieVersion: 1,
+			created: aDate,
+			lastUpdated: aDate,
+			cmpId: 1,
+			cmpVersion: 1,
+			consentScreen: 1,
+			consentLanguage: 'EN',
+			vendorListVersion: 1,
+			purposeIdBitString: '000000001010101010001100',
+			maxVendorId: 5,
+			isRange: false,
+			vendorIdBitString: '10011',
+		};
+
+		const bitString = encodeVendorCookieValue(consentData);
+		const decoded = decodeVendorCookieValue(bitString);
+
+		expect(decoded).to.deep.equal(consentData);
+	});
+
+
+	it('encodes and decodes the publisher cookie value without ranges back to original value', () => {
+
+		const aDate = new Date('2018-07-15 PDT');
+
+		const consentData = {
+			cookieVersion: 1,
+			created: aDate,
+			lastUpdated: aDate,
+			cmpId: 1,
+			cmpVersion: 0,
+			consentScreen: 0,
+			consentLanguage: 'AA',
+			vendorListVersion: 1,
+			publisherPurposeVersion: 1,
+			numCustomPurposes: 4,
+			standardPurposeIdBitString: '000000001010101010001100',
+			customPurposeIdBitString: '1011',
+		};
+
+		const bitString = encodePublisherCookieValue(consentData);
+		const decoded = decodePublisherCookieValue(bitString);
+
+		expect(decoded).to.deep.equal(consentData);
 	});
 });
